requirements.txt
reflex==0.6.3
langchain
langchain_core
langchain_google_genai
python-dotenv
langserve[all]
uvicorn
fastapi